+++ android/src/main/kotlin/com/xraph/plugin/flutter_unity_widget/FlutterUnityWidgetBuilder.kt
@@ -8,7 +8,7 @@ class FlutterUnityWidgetBuilder : FlutterUnityWidgetOptionsSink {
 
     fun build(
             id: Int,
-            context: Context,
+            context: Context?,
             binaryMessenger: BinaryMessenger,
             lifecycle: LifecycleProvider
     ): FlutterUnityWidgetController {