+++ example/lib/screens/api_screen.dart
@@ -43,7 +43,7 @@ class _ApiScreenState extends State<ApiScreen> {
                 onUnityMessage: onUnityMessage,
                 onUnitySceneLoaded: onUnitySceneLoaded,
                 fullscreen: false,
-                useAndroidViewSurface: false,
+                useAndroidViewSurface: true,
                 webUrl: 'http://localhost:6080/',
               ),
             ),