+++ example/lib/screens/loader_screen.dart
@@ -35,6 +35,7 @@ class _LoaderScreenState extends State<LoaderScreen> {
               onUnityCreated: onUnityCreated,
               onUnityMessage: onUnityMessage,
               webUrl: 'http://localhost:6080/',
+              useAndroidViewSurface: true,
             ),
             Positioned(
               bottom: 20,