+++ example/lib/screens/no_interaction_screen.dart
@@ -45,6 +45,7 @@ class _NoInteractionScreenState extends State<NoInteractionScreen> {
                 onUnityMessage: onUnityMessage,
                 onUnitySceneLoaded: onUnitySceneLoaded,
                 webUrl: 'http://localhost:6080/',
+                useAndroidViewSurface: true,
               ),
             ],
           )),