+++ lib/src/io/device_method.dart
@@ -176,6 +176,8 @@ class MethodChannelUnityWidget extends UnityWidgetPlatform {
           gestureRecognizers: gestureRecognizers,
           creationParams: creationParams,
           creationParamsCodec: const StandardMessageCodec(),
+          hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+          layoutDirection: TextDirection.ltr,
         );
       }
 
@@ -193,8 +195,8 @@ class MethodChannelUnityWidget extends UnityWidgetPlatform {
           );
         },
         onCreatePlatformView: (PlatformViewCreationParams params) {
-          final SurfaceAndroidViewController controller =
-              PlatformViewsService.initSurfaceAndroidView(
+          final ExpensiveAndroidViewController controller =
+              PlatformViewsService.initExpensiveAndroidView(
             id: params.id,
             viewType: _viewType,
             layoutDirection: TextDirection.ltr,